Version 4.181
SoNode uniqueId must be a long 64bits

+++ Koin3D/jscenegraph/src/jscenegraph/coin3d/inventor/elements/SoMultiTextureCoordinateElement.java
@@ -33,7 +33,7 @@ public class SoMultiTextureCoordinateElement extends SoElement {
 	
 	public static class UnitData implements Mutable {
 		
-	    public int nodeid;
+	    public long nodeid;
 	    public CoordType whatKind;
 	    public SoTextureCoordinateFunctionCB funcCB;
 	    public Object funcCBData;
